requirements.txt
ollama
requests
openai-whisper
pyaudio
edge-tts
pygame